Ensure the immutable validator allows deleted sync documents to be reinstated
Previously, because an immutable property on a deleted document is undefined, the validator would consider a new revision that tries to assign a value other than undefined to the property of the reinstated document to be invalid. Now, the immutable validator is simply not applied in the case that the stored document was deleted.