Test range validation for invalid datetime format
Ensures that, if a client attempts to save a date/time property that is not in a valid ISO 8601 format, it will not be considered out of range. The value will of course still fail validation because it does not match the regex pattern that is built in to the "datetime" type.

Manually verified that date/time range parsing behaves as expected in a live Sync Gateway bucket with all of the examples from these test cases.