Immutable object properties can change from null to undefined
Null and undefined are now treated as effectively equal when determining if a document or object property has been modified. However, previously they were not because the old and new objects' property counts would differ if a property was null in one and undefined in the other. Now the property counts don't affect the test.