Issue #29: Constraint to enforce immutability only on items that have values
Unlike the immutable parameter, the new immutableIfSet parameter only prevents modification to a property or element if its value is neither null nor undefined. This allows a document to be created with a property or element that is initially undefined or null and that can later be set with a value that cannot be modified. For example, a property that indicates when a message was first read may be undefined when the message is first created and then set to a date/time when the recipient actually views it, but never changed after that.