Issue #39: Improve message for mismatched lists of channels and error messages
Previously, when a document was created with one or more channels that weren't expected or a document validation failure included one or more error messages that weren't expected, the test helper would display an unhelpful message along the lines of "expected X to be Y". Now the contents of the two lists in each case are compared in a more dynamic fashion that highlights which element, if any, is incorrect/missing.