Issue #28: Parameter to allow unknown properties in documents and objects
When enabled, reduces the burden of server-side validation of documents and objects that are nested within them. When the parameter is applied to a document or object, it is not applied recursively. So if, for example, there is an object nested within a document that allows unknown properties, the object will continue to reject unknown properties by default.