Issue #22: Cascade authorization types
Instead of requiring that the user have at least one of the specified channels AND at least one of the roles when both channels and roles are defined for a document type, require that the user have one of the specified channels OR roles. This way is more intuitive and useful.