Issue #73: Implicitly allow type property when using the simple type filter
The `simpleTypeFilter` is a convenience that specifies that a document's type is defined by its "type" property. For the sake of convenience, if a document definition uses a `simpleTypeFilter` but does not explicitly include a property validator for its "type" property, automatically allow the "type" property and validate it with the built in `typeIdValidator`.