Issue #116: Avoid character sequences used as special replacement patterns
Previously, sync functions generated by synctos could not be loaded by the Sync Gateway admin UI's editor view due to a defect in the way that project injects a sync function's contents into the page. Specifically, the admin UI uses the JavaScript `String.replace` function that allows special replacement patterns in the replacement value, one of which (`$'`) appears frequently at the end of a number of regular expressions in synctos' templates. The underlying issue should be resolved in that project; however, as an immediate workaround to the problem, double quotes are now used instead of single quotes when constructing such regular expression strings.