Issue #43: Perform validation on all sample and test document definitions
Helps to ensure that the document definitions bundled with the project and used as samples and test cases conform to specifications.

Note that the `noAuthorizationsDefinedDoc` type was removed from the document definitions for the authorization specifications since it is, by definition, an invalid document definition (i.e. it does not define `channels`, `authorizedRoles` or `authorizedUsers` properties) and thus causes the build to fail. Since it is an edge case and that case is now covered by the validation utility, it seems acceptable to remove it.