Issue #43: Remove spurious unescaping of backtick characters
The document-definitions-shell-maker module "unescaped" the character sequence "\`" when encountered in a document definitions file. However, that behaviour is only desirable when the test-helper module loads a sync function that has had such escape sequences inserted by the sync-function-loader module. The sync-function-loader module escapes backtick characters so that they do not generate syntax errors due to their use as the multiline string character in Sync Gateway configuration files. Since the document definitions that are passed to document-definitions-shell-maker have not gone through the sync-function-loader, the backticks have not been escaped and any occurrence of the character sequence "\`" should be treated literally.