Apply proper deprecation to old modules in etc directory
The test-helper and validation-error-message-formatter modules now live in the `src` directory. Apply Node.js' [deprecate](https://nodejs.org/dist/latest/docs/api/util.html#util_util_deprecate_function_string) function to the affected functions from the old versions of the modules in the `etc` directory so that a deprecation warning will be written to the console when they are used.