Linter rules for JavaScript templates
Though some of the JavaScript files in the `templates` directory are by definition invalid/incomplete when in their raw state, it is still useful to specify JSHint rules; editors that lint on the fly on open documents will no longer display a litany of errors/warnings from the linter for variables and functions that do not exist in the template itself.

Also took the liberty of adding a name to the top level function in each of the various JavaScript templates.