Issue #190: Remove reference to JSON object from generated sync functions
Previously the `mustEqual` and `mustEqualStrict` constraints used `JSON.stringify` when formatting an error message indicating a constraint violation; however, since the JavaScript engine used by Sync Gateway does not support the JSON object, this resulted in JavaScript errors at runtime if a client tried to store a document with a value that violated the constraint. Removed references to the JSON object from the sync function templates and replaced it with a polyfill (`json-stringify-module.js`) that replicates the behaviour of `JSON.stringify`.