Issue #207: Ignore all internal document properties
During execution of a sync function ignore any document property key that starts with an underscore if it is at the top level of the document. Sync Gateway does not allow top-level document property keys to start with an underscore because it reserves that notation to indicate internal document properties.