Issue #185: Set minimum supported Node.js version
WARNING: Breaking change.

Updated the project's package configuration to indicate that only Node.js versions 8.9.0 or higher may be used. Technically earlier versions may work now and in the future, but they are not supported. Also increased the version of the ECMAScript specification that Node.js code supports from ES5 to ES2015/ES6.