Issue #230: Fix file path resolution for test and validation environment templates
Previously the test-environment-maker and validation-environment-maker modules used a relative file path to reference their respective template files. However, this is invalid when the module is run from a dependent project since the current working directory will be the root of _that_ project rather than the root of synctos. Now it uses Node.js' built in path module to resolve the template file paths correctly.