Issue #35: Escape backticks when enclosing a sync function in a JSON string
The backtick character has a special significance in a Sync Gateway configuration file: it is used to denote the beginning and end of a multi-line string (a feature that is not standard JSON). Its presence, even when embedded within a standard double-quoted JSON string, can confuse the parser. It is therefore necessary for the `make-sync-function` script to explicitly escape any backtick that it encounters, even when the output should be enclosed in a single-line JSON string.

Manually verified the correct behaviour in Sync Gateway after applying this change.