Execute npm install to update lockfile for Greenkeeper
Travis CI uses `npm ci` by default to install Node.js dependencies when it detects npm 5.7.0 or higher. However, `npm ci` does not update the lockfile, which means that the `package.json` and `package-lock.json` versions will be out of sync when Greenkeeper updates a package's version in `package.json` and builds fail. This change ensures that Travis CI will execute `npm install` to fetch and install the project's dependencies and also to update the `package-lock.json` file.