Issue #323: Allow validation to be skipped when an item has a strictly equal value
Introduces the `skipValidationWhenValueUnchangedStrict` constraint for all validation types. When enabled, validation is not performed on the property or element when its value is _strictly_ equal to the value from the previous revision so that, even if the value no longer satisfies the item constraints, it will be allowed. The result is that, even if the property or element no longer satisfies the item constraint(s), its value will be allowed if it has not changed.