Issue #338: Deprecated mustNotBeMissing and mustNotBeNull constraints
The `mustNotBeMissing` and `mustNotBeNull` constraints do not behave as expected since, when Sync Gateway encounters `null` values from a JSON document into an object to pass to the sync function, it removes the properties altogether. The document validation script (`validate-document-definitions`) now fails if it encounters either constraint in document definitions and prints a deprecation warning. Furthermore, the `mustNotBeNull` constraint now fails if the corresponding property value is missing/`undefined` (since `null` values become `undefined` when converted to the object that's passed to a sync function).